backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
motor>=3.1.2
pymongo>=4.3.3
python-dotenv>=1.0.0
passlib>=1.7.4
pyjwt>=2.6.0
bcrypt>=4.1.0
ortools>=9.6.2534
pandas>=2.0.0
openpyxl>=3.1.2
weasyprint>=60.0
python-multipart>=0.0.6
email-validator>=2.0.0
google-generativeai>=0.3.0
reportlab>=4.0.0
protobuf>=5.29.5
pydantic-settings>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
python-jose[cryptography]>=3.3.0
